add a reader for Yaml Documents
